--- core/imports/exports/readme.md
@@ -1,3 +1,0 @@
-## Exports
-
-Exports currently provided by the emulator itself.
