+++ modules/external/libSceGraphicsDriver/entry.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "core/imports/exports/pm4_custom.h"
+#include "pm4_custom.h"
 #include "core/kernel/eventqueue_types.h"
 #include "core/memory/memory.h"
 #include "core/timer/timer.h"
